// web/src/api/generated/models/adminListPartners200PartnersItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReloadCars API
 * API do app de pontos de recarga para veículos elétricos.
 * OpenAPI spec version: 1.0.0
 */

export type AdminListPartners200PartnersItem = {
  id: string
  name: string
  /** @nullable */
  document: string | null
  /** @nullable */
  email: string | null
  /** @nullable */
  phone: string | null
  active: boolean
  createdAt: string
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  chargePointCount: number
}